Update docstrings to reflect extra options list.
repo_extract_extra_opts may be a list of str; however docstrings only
reflected a string. In a case of multiple extra options, this caused
repo to fail as extra_opts are passed via command, which causes
ansible.builtin.unarchive to fail. A string with additional arguments
worked by happenstance as it was cast by the encapsulated task without
issue.

Changed:
* repo_extract_extra_opts - Update docstring to reflect accepted
  datatype. Existing usage of a singular string for a single option is
  still valid; documentation reflects multiple args.